Cleaned the floors for more NUGGET DIPPING ACTION
added an options nug but it doesnt work. basically I want it to reset
the score, and mute sounds/music.

also I made it harder the more you dunk! (well its based on score right
now, but it would be neat to count dunks and change it based on that!)